refactor: factory에서 API 설정 주입 방식 변경 - AuthManager 생성 시 API 설정을 외부에서 주입받도록 수정 - createAuthManagerLegacy 함수 추가로 기존 방식과의 호환성 유지
